correction
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/LHCbDIRAC/trunk@9493 4525493e-7705-40b1-a816-d608a930855b

+++ LHCbDIRAC/BookkeepingSystem/Client/BookkeepingClient.py
@@ -74,7 +74,7 @@ class BookkeepingClient:
   
   #############################################################################  
   def getProductions(self, configName, configversion, eventTypeId):
-    result = self.server.getProductions(configName, configversion, eventTypeId)
+    result = self.server.getProductions(configName, configversion, long(eventTypeId))
     return result
    
   #############################################################################
